tweak styles further, pull out component
