+++ frontend/src/components/index.ts
@@ -1,1 +1,2 @@
 export { default as Layout } from './Layout'
+export { default as InsetDrawer } from './InsetDrawer'
