+++ frontend/src/pages/Locations/index.tsx
@@ -1,14 +1,15 @@
 import * as React from 'react';
 import Map from './Map'
 import LocationsList from './LocationsList'
-import FilterDrawer from './FilterDrawer';
-import { Grid } from '@mui/material';
+import Filters from './Filters';
 import { styled } from '@mui/material/styles';
 import { LatLngTuple } from 'leaflet';
+import { InsetDrawer } from '../../components';
 
 // the hardcoded dining locations JSON file is temporary
 // @TODO: move these restaurant locations to the DB
 import dining from './dining.json'
+
 const locations = dining.features.map((f) => {
   return {
     id: parseInt(f.properties.ID, 10),
@@ -19,28 +20,30 @@ const locations = dining.features.map((f) => {
   }
 })
 
+// rework to make map a percentage of screen
 const FILTER_DRAWER_WIDTH = 225
+const MAP_DRAWER_WIDTH = '40%'
 
 const LocationListContainer = styled('div')(({ theme }) => ({
   marginLeft: FILTER_DRAWER_WIDTH,
+  marginRight: MAP_DRAWER_WIDTH,
   padding: theme.spacing(2),
-  overflowY: 'auto'
 }));
 
 
 export default function Locations() {
   
   return (
-    <Grid container sx={{height: '100%'}}>
-      <Grid item xs={12} sm={7} sx={{position: 'relative', height: '100%'}}>
-        <FilterDrawer drawerWidth={FILTER_DRAWER_WIDTH}/>
-        <LocationListContainer>
-          <LocationsList locations={locations} />
-        </LocationListContainer>
-      </Grid>
-      <Grid item xs={12} sm={5} sx={{height: 'calc(100vh - 88px)'}}>
+    <div>
+      <InsetDrawer anchor="left" width={FILTER_DRAWER_WIDTH}>
+        <Filters />
+      </InsetDrawer>
+      <LocationListContainer>
+        <LocationsList locations={locations} />
+      </LocationListContainer>
+      <InsetDrawer anchor="right" width={MAP_DRAWER_WIDTH}>
         <Map locations={locations} />  
-      </Grid>
-    </Grid>
+      </InsetDrawer>
+    </div>
   );
 };
